requirements.txt
adversarial-robustness-toolbox==1.5.1
art==5.1
cached-property==1.5.2
cachetools==4.0.0
foolbox==3.2.1
graphviz==0.16
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.5.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterhub==1.0.0
jupytext==1.4.1
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
nbconvert==5.5.0
nbformat==4.4.0
nltk==3.4.4
notebook==5.7.8
numba==0.52.0
numpy==1.17.3
opencv-python==4.4.0.44
opt-einsum==3.1.0
pamela==1.0.0
pandas==0.24.2
Pillow==5.1.0
pip-autoremove==0.9.1
plotly==3.10.0
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.9.0
tensorflow==2.3.1
tensorflow-gpu
tqdm==4.54.0
xgboost==1.2.1
